aws-braket/requirements.txt
amazon-braket-default-simulator==1.0.0.post2
amazon-braket-schemas==1.0.2.post1
amazon-braket-sdk==1.1.3
backoff==1.10.0
boltons==20.2.1
boto3==1.16.3
botocore==1.19.3
decorator==4.4.2
jmespath==0.10.0
nest-asyncio==1.4.1
networkx==2.5
numpy==1.19.2
opt-einsum==3.3.0
pydantic==1.6.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
